requirements.txt
actionlib==1.12.1
angles==1.9.12
anyio==3.6.2
apturl==0.5.2
argcomplete==1.10.3
asgiref==3.4.1
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
black==22.8.0
bondpy==1.8.5
Brlapi==0.6.6
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
ci-info==0.2.0
click==8.0.4
colorama==0.3.7
command-not-found==0.3
compressed-rtf==1.0.6
configobj==5.0.8
configparser==5.2.0
contextlib2==21.6.0
contextvars==2.4
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cryptography==2.1.4
cupshelpers==1.0
cv_bridge==1.13.1
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro-info===0.18ubuntu0.18.04.1
dlib==19.16.0
docx2txt==0.8
dynamic_reconfigure==1.6.5
ebcdic==1.1.1
etelemetry==0.2.2
extract-msg==0.28.7
fastapi==0.83.0
filelock==3.4.1
fitz==0.0.1.dev2
Flask==2.0.3
future==0.18.3
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
h11==0.13.0
httplib2==0.9.2
httptools==0.3.0
idna==3.4
image_geometry==1.13.1
IMAPClient==2.1.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
interactive-markers==1.11.5
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
joint_state_publisher==1.12.15
kdl-parser-py==1.13.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.9.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
message_filters==1.14.13
mypy-extensions==1.0.0
netifaces==0.10.4
networkx==2.5.1
nibabel==3.2.2
nipype==1.7.1
numpy==1.19.5
oauth==1.0.1
olefile==0.46
opencv-python==4.7.0.68
packaging==21.3
pandas==1.1.5
pathlib==1.0.1
pathspec==0.9.0
pdfminer.six==20191110
pexpect==4.2.1
Pillow==5.1.0
platformdirs==2.4.0
protobuf==3.0.0
prov==2.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.17
pycups==1.9.73
pydantic==1.9.2
pydot==1.4.2
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF4==1.27.0
pyRFC3339==1.0
pytesseract==0.3.8
python-apt==1.6.6
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.20.0
python-multipart==0.0.5
python-pptx==0.6.21
python_qt_binding==0.4.4
pytz==2018.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
pyxnat==1.5
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rdflib==5.0.0
reportlab==3.4.0
requests==2.27.1
requests-unixsocket==0.1.5
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.29
scipy==1.5.4
SecretStorage==2.3.1
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.16.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
starlette==0.19.1
system-service==0.3
systemd-python==234
textract==1.6.5
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
tomli==1.2.3
topic_tools==1.14.13
traits==6.4.1
typed-ast==1.5.4
typing_extensions==4.1.1
tzdata==2023.2
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.6
urllib3==1.22
usb-creator==0.3.3
uvicorn==0.16.0
uvloop==0.14.0
wadllib==1.3.2
Wand==0.6.11
watchgod==0.7
websockets==9.1
Werkzeug==2.0.3
xacro==1.13.18
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.9
zipp==3.6.0
zope.interface==4.3.2
